+++ Jantimizer/Histograms/HistogramModels/IHistogramBucket.cs
@@ -4,6 +4,6 @@
     {
         public IComparable ValueStart { get; }
         public IComparable ValueEnd { get; }
-        public int Count { get; }
+        public long Count { get; }
     }
 }
